@@ -8,19 +8,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.2.2",
-    "@angular/cdk": "^10.2.7",
-    "@angular/common": "^10.2.2",
-    "@angular/compiler": "^10.2.2",
-    "@angular/core": "^10.2.2",
-    "@angular/fire": "^6.0.4",
-    "@angular/forms": "^10.2.2",
-    "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "^10.2.2",
-    "@angular/platform-browser-dynamic": "^10.2.2",
-    "@angular/pwa": "^0.1002.0",
-    "@angular/router": "^10.2.2",
-    "@angular/service-worker": "^10.2.2",
+    "@angular/animations": "^11.0.0",
+    "@angular/cdk": "^11.0.0",
+    "@angular/common": "^11.0.0",
+    "@angular/compiler": "^11.0.0",
+    "@angular/core": "^11.0.0",
+    "@angular/fire": "^6.0.5",
+    "@angular/forms": "^11.0.0",
+    "@angular/material": "^11.0.0",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@angular/pwa": "^0.1100.0",
+    "@angular/router": "^11.0.0",
+    "@angular/service-worker": "^11.0.0",
     "@ngxs-labs/data": "^4.1.1",
     "@ngxs/storage-plugin": "^3.7.0",
     "@ngxs/store": "^3.7.0",
@@ -30,15 +30,15 @@
     "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1002.0",
-    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular-devkit/architect": "^0.1100.0",
+    "@angular-devkit/build-angular": "^0.1100.0",
     "@angular/cli": "^11.0.0",
-    "@angular/compiler-cli": "^10.2.2",
-    "@angular/language-service": "^10.2.2",
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/language-service": "^11.0.0",
     "@types/hammerjs": "^2.0.36",
     "@types/node": "^14.14.7",
     "codelyzer": "^6.0.1",
     "tslint": "^6.1.3",
-    "typescript": "^3.9.7"
+    "typescript": "^4.0.5"
   }
 }
